+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Change Log
 
-## 1.2.2-alpha.7 - (unreleased)
+## 1.2.2-alpha.7 - 2026.01.24
 
 ### Added
 
 - Added support for extracting JS files from client-side paths (HTML) in Next.js
 
 ### Changed
+
+- Fixed recursive script tag parsing
 
 ### Fixed